added a clean_soup function so that 'get_pv_systems_for_country' function works again with soup now being scraped from site
